to 2019

+++ com.unity.render-pipelines.universal/Shaders/2D/Include/CombinedShapeLightShared.hlsl
@@ -17,7 +17,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter0))
     {
-        half4 processedMask = (1 - _ShapeLightInvertedFilter0) * mask + _ShapeLightInvertedFilter0 * (1 - mask);
+        float4 processedMask = (1 - _ShapeLightInvertedFilter0) * mask + _ShapeLightInvertedFilter0 * (1 - mask);
         shapeLight0 *= dot(processedMask, _ShapeLightMaskFilter0);
     }
 
@@ -33,7 +33,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter1))
     {
-        half4 processedMask = (1 - _ShapeLightInvertedFilter1) * mask + _ShapeLightInvertedFilter1 * (1 - mask);
+        float4 processedMask = (1 - _ShapeLightInvertedFilter1) * mask + _ShapeLightInvertedFilter1 * (1 - mask);
         shapeLight1 *= dot(processedMask, _ShapeLightMaskFilter1);
     }
 
@@ -49,7 +49,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter2))
     {
-        half4 processedMask = (1 - _ShapeLightInvertedFilter2) * mask + _ShapeLightInvertedFilter2 * (1 - mask);
+        float4 processedMask = (1 - _ShapeLightInvertedFilter2) * mask + _ShapeLightInvertedFilter2 * (1 - mask);
         shapeLight2 *= dot(processedMask, _ShapeLightMaskFilter2);
     }
 
@@ -65,7 +65,7 @@ half4 CombinedShapeLightShared(half4 color, half4 mask, half2 lightingUV)
 
     if (any(_ShapeLightMaskFilter3))
     {
-        half4 processedMask = (1 - _ShapeLightInvertedFilter3) * mask + _ShapeLightInvertedFilter3 * (1 - mask);
+        float4 processedMask = (1 - _ShapeLightInvertedFilter3) * mask + _ShapeLightInvertedFilter3 * (1 - mask);
         shapeLight3 *= dot(processedMask, _ShapeLightMaskFilter3);
     }
